ReadEasy.MainDriver/requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1654189308200/work
aiohttp @ file:///Users/runner/miniforge3/conda-bld/aiohttp_1649013197268/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1636093929600/work
astunparse @ file:///home/conda/feedstock_root/build_artifacts/astunparse_1610696312422/work
async-timeout @ file:///home/conda/feedstock_root/build_artifacts/async-timeout_1640026696943/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1640799537051/work
bidict @ file:///home/conda/feedstock_root/build_artifacts/bidict_1648087833370/work
blinker==1.4
brotlipy @ file:///Users/runner/miniforge3/conda-bld/brotlipy_1648854210220/work
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1633010882559/work
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_83242e7e-f82d-4a71-8ef2-9d71d212d249gu_wxmeq/croots/recipe/certifi_1655968827803/work/certifi
cffi @ file:///Users/builder/ci_310/cffi_1643393868399/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1655906222726/work
click @ file:///opt/concourse/worker/volumes/live/2d66025a-4d79-47c4-43be-6220928b6c82/volume/click_1646056610594/work
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography_1657174324676/work
docopt==0.6.2
flatbuffers @ file:///home/conda/feedstock_root/build_artifacts/python-flatbuffers_1620718971772/work
frozenlist @ file:///Users/runner/miniforge3/conda-bld/frozenlist_1648771715263/work
gast @ file:///home/conda/feedstock_root/build_artifacts/gast_1596839682936/work
google-auth @ file:///home/conda/feedstock_root/build_artifacts/google-auth_1629296548061/work
google-auth-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/google-auth-oauthlib_1630497468950/work
google-pasta==0.2.0
grpcio @ file:///Users/runner/miniforge3/conda-bld/grpcio_1637189415781/work
h5py @ file:///Users/builder/ci_310/h5py_1642617990072/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1642433548627/work
importlib-metadata @ file:///Users/runner/miniforge3/conda-bld/importlib-metadata_1653252891204/work
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
keras @ file:///home/conda/feedstock_root/build_artifacts/keras_1637872120405/work/keras-2.7.0-py2.py3-none-any.whl
Keras-Preprocessing @ file:///home/conda/feedstock_root/build_artifacts/keras-preprocessing_1610713559828/work
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1651821407140/work
mkl-fft==1.3.1
mkl-random @ file:///opt/concourse/worker/volumes/live/a9eca864-069c-4240-73c8-4e51b036a10d/volume/mkl_random_1639994047876/work
mkl-service==2.4.0
multidict @ file:///Users/runner/miniforge3/conda-bld/multidict_1648882466411/work
nltk @ file:///opt/conda/conda-bld/nltk_1645628263994/work
num2words @ file:///home/conda/feedstock_root/build_artifacts/num2words_1626355108962/work
numpy @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_e3f85f54-0572-4c0e-b190-4bc4766fc3fenewxarhq/croots/recipe/numpy_and_numpy_base_1652801682879/work
oauthlib @ file:///home/conda/feedstock_root/build_artifacts/oauthlib_1643507977997/work
opt-einsum @ file:///home/conda/feedstock_root/build_artifacts/opt_einsum_1617859230218/work
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyAudio==0.2.11
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1652398519695/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1643496850550/work
PySocks @ file:///Users/runner/miniforge3/conda-bld/pysocks_1648857336495/work
python-engineio @ file:///home/conda/feedstock_root/build_artifacts/python-engineio_1656977683910/work
python-Levenshtein @ file:///Users/runner/miniforge3/conda-bld/python-levenshtein_1649606984621/work
python-socketio @ file:///home/conda/feedstock_root/build_artifacts/python-socketio_1656977625393/work
pyu2f @ file:///home/conda/feedstock_root/build_artifacts/pyu2f_1604248910016/work
regex @ file:///opt/concourse/worker/volumes/live/63aed588-2398-47ee-5ad3-e682b77e31f2/volume/regex_1648447727979/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1656534056640/work
requests-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/requests-oauthlib_1643557462909/work
rsa @ file:///home/conda/feedstock_root/build_artifacts/rsa_1637781155505/work
scipy @ file:///Users/builder/ci_310/scipy_1642604663510/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
SpeechRecognition @ file:///home/conda/feedstock_root/build_artifacts/speechrecognition_1639389450396/work
tensorboard @ file:///home/conda/feedstock_root/build_artifacts/tensorboard_1629677129676/work/tensorboard-2.6.0-py3-none-any.whl
tensorboard-data-server @ file:///Users/runner/miniforge3/conda-bld/tensorboard-data-server_1649932892572/work/tensorboard_data_server-0.6.0-py3-none-macosx_10_9_x86_64.whl
tensorboard-plugin-wit @ file:///home/conda/feedstock_root/build_artifacts/tensorboard-plugin-wit_1641458951060/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorflow @ file:///Users/uwe/Development/conda-forge/tensorflow-feedstock/tensorflow-feedstock-osx_64_numpy1.21python3.10.____cpython/miniforge3/conda-bld/tensorflow-split_1644027665818/work/tensorflow_pkg/tensorflow-2.7.0-cp310-cp310-macosx_10_9_x86_64.whl
tensorflow-estimator @ file:///Users/uwe/Development/conda-forge/tensorflow-feedstock/tensorflow-feedstock-osx_64_numpy1.21python3.10.____cpython/miniforge3/conda-bld/tensorflow-split_1644027665818/work/tensorflow-estimator/wheel_dir/tensorflow_estimator-2.7.0-py2.py3-none-any.whl
tensorflow-hub @ file:///home/conda/feedstock_root/build_artifacts/tensorflow-hub_1618768305670/work/wheel_dir/tensorflow_hub-0.12.0-py2.py3-none-any.whl
termcolor @ file:///home/conda/feedstock_root/build_artifacts/termcolor_1657118200573/work
tqdm @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_6cb3db64-b7e1-4db2-ab0b-f944920e5367lekntvc5/croots/recipe/tqdm_1650891088614/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1656706066251/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1657224465922/work
websocket-client==1.3.3
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1651670883478/work
wrapt @ file:///Users/runner/miniforge3/conda-bld/wrapt_1651495286247/work
yarl @ file:///Users/runner/miniforge3/conda-bld/yarl_1648966580272/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1649012893348/work
